requirements.txt
fastapi
uvicorn
numpy
scikit-learn
sentence-transformers
faiss-cpu
pandas